+++ apps/evm/app/layout.tsx
@@ -4,8 +4,8 @@ import { SanctionedAddressDialog } from '@sushiswap/wagmi/future/components/Sanc
 import type { Metadata } from 'next'
 import { Inter, Roboto_Mono } from 'next/font/google'
 import React from 'react'
-
 import { Providers } from './providers'
+import { Trackers } from './trackers'
 
 const inter = Inter({
   subsets: ['latin'],
@@ -43,6 +43,7 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
           <SanctionedAddressDialog />
           {children}
         </Providers>
+        <Trackers />
       </body>
     </html>
   )